requirements.txt
anyio==3.6.1
asgiref==3.5.2
bcrypt==4.0.0
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.4
defusedxml==0.7.1
distlib==0.3.6
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.85.0
filelock==3.8.0
gunicorn==20.1.0
h11==0.13.0
httptools==0.5.0
identify==2.5.3
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
nodeenv==1.7.0
oauthlib==3.2.0
orjson==3.8.0
passlib==1.7.4
platformdirs==2.5.2
pre-commit==2.20.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
PyJWT==2.4.0
python-decouple==3.6
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlparse==0.4.2
starlette==0.20.4
toml==0.10.2
typing-extensions==4.3.0
ujson==5.4.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
virtualenv==20.16.4
watchfiles==0.17.0
websockets==10.3
